Media: fix remove media, correctly remove it from server and pagecontent

+++ dao/media_page.js
@@ -40,20 +40,23 @@ export async function selectMediaPage(media_id, page_id) {
 }
 
 export async function deleteMediaPage(media_id, page_id) {
-    // const res = await query(
-    //     `
-    //     DELETE FROM media_page
-    //     WHERE media_id = ? AND page_id = ?
-    // `,
-    //     [media_id, page_id],
-    // );
-
-    // return res.affectedRows;
-    const res = await prisma.media_page.delete({
-        where: {
-            media_id: parseInt(media_id),
-            page_id,
-        },
-    });
-    return res;
+    try {
+        const mediaPage = await prisma.media_page.findMany({
+            where: {
+                media_id: parseInt(media_id),
+                page_id: parseInt(page_id),
+            },
+        });
+        console.log({mediaPage});
+        if (!mediaPage) return null;
+        const res = await prisma.media_page.delete({
+            where: {
+                id: mediaPage[0].id,
+            },
+        });
+        return res;
+    } catch (error) {
+        console.log({error});
+        return null;
+    }
 }
